# Speed-up casts to FP8