@drisspg 2023-10-02
Ran `python benchmarks/bench_linear_float8.py --use_ts -o benchmarks/data/float8_sweep_cast_update` before and after this pr:
| name      | shape               | ref_dtype      |   pt_fp8_time_sec |   pt_fp8_time_sec_updated |   % Improvement |
|:----------|:--------------------|:---------------|------------------:|--------------------------:|----------------:|
| attn.wqkv | (16384, 8192, 1280) | torch.bfloat16 |        0.00416412 |                0.00390124 |         6.31303 |
| attn.w0   | (16384, 1024, 8192) | torch.bfloat16 |        0.00361837 |                0.00335387 |         7.30976 |
| ffn.w13   | (16384, 8192, 7168) | torch.bfloat16 |        0.0107657  |                0.0102925  |         4.39569 |
| ffn.w2    | (16384, 3584, 8192) | torch.bfloat16 |        0.00660132 |                0.00621062 |         5.91851 |
| attn.wqkv | (16384, 8192, 1280) | torch.float16  |        0.00423888 |                0.00398655 |         5.95275 |
| attn.w0   | (16384, 1024, 8192) | torch.float16  |        0.00367196 |                0.0034114  |         7.09605 |
| ffn.w13   | (16384, 8192, 7168) | torch.float16  |        0.0107916  |                0.010287   |         4.67511 |
| ffn.w2    | (16384, 3584, 8192) | torch.float16  |        0.00662924 |                0.00625864 |         5.59038 |


Looks like this is producing very meaningful speedups!